pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "poseidon"
version = "0.1.0"
description = "Scientific data analysis and visualization tool"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "python-dotenv>=1.0.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.1.19",
    "fastapi>=0.116.0",
    "uvicorn>=0.35.0",
    "redis>=6.4.0",
    "pyyaml>=6.0.2",
    "plotly>=6.5.2",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
]

[tool.setuptools.packages.find]
where = ["src"]
